# sync-bridge-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Transactional data synchronization API for SQLite target databases.",
        "title": "Sync Bridge Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/sync": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Syncs data payload items (customers, products, orders, employees) transactionally. If one fails, the target database modifications roll back, but the execution log is committed to sync_history with 'FAILED' status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "Synchronize data payload",
                "parameters": [
                    {
                        "description": "Data Payload to synchronize",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SyncRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SyncResponse"
                        }
                    },
                    "400": {
                        "description": "Validation failed or invalid JSON",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "409": {
                        "description": "Data Integrity Violation (Conflict)",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/sync-history": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of sync history logs, optionally filtered by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "List sync history",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (SUCCESSFUL, FAILED, PENDING_RETRY, INVALID)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListHistoryResponse"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/sync-history/retry/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Triggers a retry operation on a failed sync history by resetting its status to PENDING_RETRY.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "Retry failed sync history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync History ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RetryHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid sync history ID or only failed syncs can be retried",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Sync history not found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/sync-history/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetches a single sync history record by its database ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "Get sync history details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync History ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid sync history ID",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Sync history not found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a sync history log entry from the database by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "Delete sync history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync History ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid sync history ID",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Sync history not found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/sync/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Emits aggregated counts of sync histories grouped by status type (SUCCESSFUL, FAILED, PENDING_RETRY, INVALID)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "Get sync statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.StatsResponse"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.SyncRequest": {
            "type": "object",
            "required": [
                "data",
                "model"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "object"
                    }
                },
                "model": {
                    "type": "string",
                    "enum": [
                        "customers",
                        "products",
                        "orders",
                        "employees"
                    ]
                }
            }
        },
        "errors.AppError": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "handlers.GetHistoryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.SyncHistoryRow"
                },
                "message": {
                    "type": "string",
                    "example": "Sync history retrieved successfully"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "handlers.ListHistoryData": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SyncHistoryRow"
                    }
                },
                "empty": {
                    "type": "boolean",
                    "example": false
                },
                "first": {
                    "type": "boolean",
                    "example": true
                },
                "last": {
                    "type": "boolean",
                    "example": false
                },
                "number": {
                    "type": "integer",
                    "example": 0
                },
                "numberOfElements": {
                    "type": "integer",
                    "example": 15
                },
                "pageable": {
                    "$ref": "#/definitions/handlers.PageableInfo"
                },
                "size": {
                    "type": "integer",
                    "example": 15
                },
                "totalElements": {
                    "type": "integer",
                    "example": 25
                },
                "totalPages": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "handlers.ListHistoryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.ListHistoryData"
                },
                "message": {
                    "type": "string",
                    "example": "Sync histories retrieved successfully"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "handlers.PageableInfo": {
            "type": "object",
            "properties": {
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "pageNumber": {
                    "type": "integer",
                    "example": 0
                },
                "pageSize": {
                    "type": "integer",
                    "example": 15
                },
                "paged": {
                    "type": "boolean",
                    "example": true
                },
                "sort": {
                    "$ref": "#/definitions/handlers.SortInfo"
                },
                "unpaged": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handlers.RetryHistoryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.SyncHistoryRow"
                },
                "message": {
                    "type": "string",
                    "example": "Sync history will be retried"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "handlers.SortInfo": {
            "type": "object",
            "properties": {
                "empty": {
                    "type": "boolean",
                    "example": false
                },
                "sorted": {
                    "type": "boolean",
                    "example": true
                },
                "unsorted": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handlers.StatsData": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "integer",
                    "example": 2
                },
                "invalid": {
                    "type": "integer",
                    "example": 0
                },
                "pending_retry": {
                    "type": "integer",
                    "example": 1
                },
                "successful": {
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "type": "integer",
                    "example": 13
                }
            }
        },
        "handlers.StatsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.StatsData"
                },
                "message": {
                    "type": "string",
                    "example": "Stats retrieved successfully"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "handlers.SyncHistoryRow": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-06-03T07:46:58.000Z"
                },
                "failureReason": {
                    "type": "string",
                    "example": "Duplicate entry: field 'EMAIL' already exists"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "payload": {
                    "type": "string",
                    "example": "[{\"email\":\"john.doe@example.com\"}]"
                },
                "retries": {
                    "type": "integer",
                    "example": 0
                },
                "status": {
                    "type": "string",
                    "example": "FAILED"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-06-03T07:46:58.000Z"
                }
            }
        },
        "handlers.SyncResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.SyncResponseData"
                },
                "message": {
                    "type": "string",
                    "example": "Sync completed successfully"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "handlers.SyncResponseData": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SyncResult"
                    }
                }
            }
        },
        "handlers.SyncResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 101
                },
                "status": {
                    "type": "string",
                    "example": "created"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-auth-token",
            "in": "header"
        }
    }
}